fix: remove orphaned types.rs duplicate module (issue #1140)
- Remove pub mod types; from lib.rs (types.rs was an 876-line orphaned
  duplicate of types already defined in lib.rs)
- Move RecipientShareDelegated struct from types.rs into lib.rs
  (only type actually used from the module path)
- Fix integration_suite.rs to use fluxora_stream::StreamKind instead
  of fluxora_stream::types::StreamKind
- Delete contracts/stream/src/types.rs

Closes #1140
